requirements.txt
python-telegram-bot[job-queue]>=20.0
reportlab>=3.6.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
qrcode[pil]>=7.4.2
pillow>=10.0.0
